requirements.txt
pandas
vnstock==3.2.3
python-dotenv
google-cloud-storage
pyarrow
fastparquet
pyspark